python/requirements.txt
apscheduler==3.10.4
beautifulsoup4==4.12.3
fastapi==0.115.2
grpcio-status==1.67.0
langchain-community==0.3.8
langchain-google-vertexai==2.0.5
langchain-openai==0.2.6
langgraph==0.2.39
pandas_ta==0.3.14b0
pip==24.2
playwright==1.47.0
python-dotenv==1.0.1
pyupbit==0.2.34
schedule==1.2.2
setuptools==75.2.0
uvicorn==0.32.0